// src/shared/models/user/model.ts
export class User {
    id:number=0;
    firstname: String='';
    lastname: String='';
    email: String='';
    address: String='';
   

    // constructor(id: number, name: String, username: String, email: String, address: String) {
    //     this.id = id
    //     this.name = name;
    //     this.username= username;
    //     this.email = email;
    //     this.address = address
    // }
  }
  